offer referal pending

+++ adm/models.py
@@ -35,7 +35,7 @@ class AdmProducts(models.Model):
     offer_price = models.DecimalField(max_digits=10, decimal_places=2)
     price = models.DecimalField(max_digits=10, decimal_places=2)
     offer_type = models.CharField(max_length=50, null=True, blank=True)
-    discount_percentage = models.DecimalField(max_digits=5, decimal_places=2) 
+    discount_percentage = models.IntegerField() 
     description = models.CharField(max_length=500)
     status = models.CharField(
         max_length=20,
